+++ driven/storage/adapter.go
@@ -603,6 +603,18 @@ func (sa *Adapter) SaveUser(user *model.User) error {
 	return nil
 }
 
+//GetCovid19Configs gives all covid 19 configs
+func (sa *Adapter) GetCovid19Configs() ([]model.COVID19Config, error) {
+	filter := bson.D{}
+	var result []model.COVID19Config
+	err := sa.db.configs.Find(filter, &result, nil)
+	if err != nil {
+		return nil, err
+	}
+	return result, nil
+
+}
+
 //ReadCovid19Config reads the covid19 configuration from the storage
 func (sa *Adapter) ReadCovid19Config() (*model.COVID19Config, error) {
 	filter := bson.D{primitive.E{Key: "name", Value: "covid19"}}